requirements.txt
Flask==3.1.2
lark-oapi @ git+https://github.com/SeanDictionary/oapi-sdk-python.git --no-build-isolation
prompt_toolkit==3.0.52
pillow==12.0.0